requirements.txt
streamlit
pandas
requests
python-dotenv
Pillow
google-generativeai
